feat: allow to move tasks between states
Implements a simple non-reactive way to change the status of
a task by moving it between kanban columns via drap and drop.
The option is server-based, so the tasks are reloaded after each
change. In addition, it is not reactive, other clients must also
reload for the change to become visible.